Change the CountMetric.recordCount method to take a Long rather than an Int as this is how it is stored internally and exposed elsewhere

+++ management/src/main/scala/com/gu/management/metrics.scala
@@ -64,7 +64,7 @@ class CountMetric(
   val `type`: String = "counter"
 
   private val _count = new AtomicLong()
-  def recordCount(count: Int): Long = _count.addAndGet(count)
+  def recordCount(count: Long): Long = _count.addAndGet(count)
   def increment(): Long = recordCount(1)
 
   def count = _count.get
